Proposed Seat NonBid code for responses rejected for privacy reasons
This is being driven by DSA and validations that Prebid Server is making on the bid responses. e.g. if the request DSA field states "bid responses without DSA object will not be accepted", a reject reason would be helpful.

Mirroring the "Request Rejected - Privacy" that was added due to GDPR, this PR proposes "Response Rejected - Privacy" for the response side of the equation.